Add CI archive action and a task to build native libs
The Archive action creates a maven artifact of the library and
make it downloadable, so that it can be published manually on
a maven repository. The action doesn't handle publishing.

This action is triggered manually. The archive's version will be
the current project's version.

This action replaces the snapshot and the release actions.

Also added a task in `native` module to build all native libs
and updated the README to use this task.