Upgrade to kotlin 1.9
Gradle is upgraded to 7.5, AGP to 7.4.2.

Ktor is upgraded to 2.3.7, and secp to a SNAPSHOT version.

The iOS source sets configuration has also been updated
to match the new KMP syntax : we support iOSX64 and
iOSArm64 out of the box.

However, iOSSimulatorArm64 is not listed as it requires
fixes in the `tor_in_thread` module first.

The native libraries have not been updated yet - but should
be afterward, in a separate commit.